Fixed issues with streamer pipeline for local files
Switched to splitmuxsink which allows rotating files based on size (or time), set the file size to 2GB. 

I also made the audio source a variable for the moment, if we had a specific audio device then we could use
`arecord -l` to determine the appropriate device number before launching the pipeline.

This really needs to switch to a python script that uses the Gstreamer python extensions to setup the pipeline.  By doing that we would be able to put the timestamp in the filename which would cause subsequent runs not to overwrite previous files.